Feature/calendar2

+++ lib/domain/repositories/calendar/events_repository.dart
@@ -1,10 +1,18 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:primary_school/domain/models/event_model/event_model.dart';
 
 class EventsRepository {
   Stream<List<EventModel>> getEventsStream() {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userId)
         .collection('calendarItems')
+        .orderBy('selectedDay')
         .snapshots()
         .map((querySnapshot) {
       return querySnapshot.docs.map(
@@ -22,23 +30,40 @@ class EventsRepository {
   }
 
   Future<void> delete({required String id}) async {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userId)
         .collection('calendarItems')
         .doc(id)
         .delete();
   }
 
   Future<void> edit(
-    Map<String, dynamic>? updatedFields,
+    String title,
+    String subtitle,
+    DateTime selectedDay,
+    DateTime selectedTime,
     String docID,
   ) async {
-    if (updatedFields == null || updatedFields.isEmpty) {
-      return;
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
     }
     await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userId)
         .collection('calendarItems')
         .doc(docID)
-        .update(updatedFields);
+        .update({
+      'title': title,
+      'subtitle': subtitle,
+      'selectedDay': selectedDay,
+      'selectedTime': selectedTime,
+    });
   }
 
   Future<void> add(
@@ -47,7 +72,15 @@ class EventsRepository {
     DateTime selectedDay,
     DateTime selectedTime,
   ) async {
-    await FirebaseFirestore.instance.collection('calendarItems').add(
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userId)
+        .collection('calendarItems')
+        .add(
       {
         'title': title,
         'subtitle': subtitle,
